+++ lib/mobility/plugins/ransack.rb
@@ -18,33 +18,6 @@ module Mobility
           names.each do |name|
             klass.ransacker(name) { backend_class.build_node(name, Mobility.locale) }
           end
-          klass.extend ClassMethods
-        end
-      end
-
-      module ClassMethods
-        def ransack(*)
-          super.extend(Search)
-        end
-      end
-
-      module Search
-        def result(opts = {})
-          sorted = sorts.inject(super) do |relation, sort|
-            predicate = ::Ransack::Visitor.new.visit_Ransack_Nodes_Sort(sort)
-            apply_mobility_scope(relation, predicate, [sort.attr_name])
-          end
-          conditions.inject(sorted) do |relation, condition|
-            apply_mobility_scope(relation, condition.arel_predicate, condition.attributes.compact.flat_map(&:name))
-          end
-        end
-
-        private
-
-        def apply_mobility_scope(relation, predicate, attributes)
-          (attributes & object.mobility_attributes).inject(relation) do |i18n_rel, attr|
-            object.mobility_backend_class(attr).apply_scope(i18n_rel, predicate)
-          end
         end
       end
     end
